gasq/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1

SQLAlchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0

pydantic==2.10.3
pydantic-settings==2.6.1

python-jose==3.3.0
passlib[bcrypt]==1.7.4

python-multipart==0.0.12
jinja2==3.1.4

httpx==0.27.2
requests==2.32.3

redis==5.2.0
celery==5.4.0

firebase-admin==6.6.0
qrcode==7.4.2
pillow==10.4.0

python-dateutil==2.9.0.post0
gunicorn==21.2.0
psycopg2-binary==2.9.9